% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akmedoids.clust.R
\name{akmedoids.clust}
\alias{akmedoids.clust}
\title{Clustering of longitudinal data}
\usage{
akmedoids.clust(traj, id_field = FALSE, method = "linear", k = c(3,6))
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observations at consecutive time steps.}

\item{id_field}{Whether the first column is a unique (\code{id}) field. Default: \code{FALSE}}

\item{method}{Initialisation strategy. Available method: \code{linear}}

\item{k}{either an exact integer number of clusters, or a vector of length two specifying the minimum and maximum numbers of clusters to be examined. The default is \code{c(3,6)}. When k is a range, the actual number of clusters is determined by Calinski-Harabatz criterion.}
}
\value{
A list containing cluster solutions for all value of k, including the solution at the optimal value of \code{k}, based on the Calinski-Harabatz criterion \code{(Calinski T, Harabasz J, 1974)}
}
\description{
This function group trajectories based on a functional definition of their long-term trends.
}
\details{
Given a list of trajectories and a functional method, this function clusters the trajectories into a k number of groups. If a vector of two numbers is given, the function determines the best solution based on the Calinski-Harabatz criterion.
}
\examples{
traj <- gm.crime.sample1
print(traj)
traj <- missingVal(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
traj <- props(traj, id_field = TRUE)
print(traj)
output <- akmedoids.clust(traj, id_field = TRUE, method = "linear", k = c(3,6))
print(output)
}
\references{
\code{Calinski T, Harabasz J (1974) A dendrite method for cluster analysis. Commun Stat 3:1-27}

\code{Christophe Genolini, Xavier Alacoque, Marianne Sentenac, Catherine Arnaud (2015). kml and kml3d: R Packages to Cluster Longitudinal Data. Journal of Statistical Software, 65(4), 1-34. URL http://www.jstatsoft.org/v65/i04/}
}
