% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaLabel.R
\name{alphaLabel}
\alias{alphaLabel}
\title{Numerics ids to alphabetical ids}
\usage{
alphaLabel(x)
}
\arguments{
\item{x}{A vector of numeric ids}
}
\description{
Function to transform a list of numeric ids to alphabetic ids
}
