% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clustr}
\alias{clustr}
\title{Sample labels of cluster groups}
\format{
A dataframe containing one variable:
\itemize{
  \item label: alphabetical label by clusters
    }
}
\usage{
clustr
}
\description{
A dataframe of alphabetical labels representing
the optimal solution of `traj` dataset based on `akClust`
function
}
\keyword{datasets}
