% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImputation.R
\name{dataImputation}
\alias{dataImputation}
\title{Data imputation for longitudinal data}
\usage{
dataImputation(traj, id_field = FALSE, method = 2, replace_with = 1, fill_zeros = FALSE)
}
\arguments{
\item{traj}{[matrix (numeric)]: longitudinal data. Each row represents an individual trajectory (of observations). The columns show the observations at consecutive time steps.}

\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.}

\item{method}{[an integer] indicating a method for calculating the missing values. Options are: \code{1: arithmetic} method, and \code{2: regression} method. The default is \code{1: arithmetic} method}

\item{replace_with}{representing the technique,}

\item{fill_zeros}{[TRUE or FALSE] whether to consider zeros \code{0} as missing values when \code{2: regression} method is used. The default is \code{FALSE}.}
}
\value{
A data.frame with missing values \code{(NA, Inf, null)} imputed according to the a specified technique.
}
\description{
This function fills any missing entries \code{(NA, Inf, null)} in a matrix or dataframe, according to a specified method. By default, \code{'0'} is considered a value.
}
\details{
Given a matrix or data.frame with some missing values indicated by \code{(NA, Inf, null)}, this function impute the missing value by using either an estimation from the corresponding rows or columns, or to use a regression method to estimate the missing values.
}
\examples{
traj <- gm.crime.sample1
 print(traj)
 dataImputation(traj, id_field = TRUE, method = 1, replace_with = 1, fill_zeros = FALSE)
}
