% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbow_point.R
\name{elbow_point}
\alias{elbow_point}
\title{Determine the elbow point on a curve}
\usage{
elbow_point(x, y)
}
\arguments{
\item{x}{vector of x coordinates of points on the curve}

\item{y}{vector of y coordinates of points on the curve}
}
\value{
indicate the optimal k value determined by
the elbow point point.
}
\description{
Given a list of x, y coordinates on a curve,
function determines the elbow point of the curve.
}
\details{
highlight the maximum curvature to identify the
elbow point (credit: 'github.com/agentlans')
}
\examples{

data(TO1Risk)

output <- akclustr(TO1Risk, id_field = FALSE,
                  method = "linear", k = c(3,20),
                  crit='Silhouette',
                  verbose = TRUE, quality_plot=FALSE)

print(output$optimal_k)

}
