% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popl.R
\docType{data}
\name{popl}
\alias{popl}
\title{sample population (denominator) data}
\format{
A matrix
}
\usage{
popl
}
\description{
simulated denominator data for two consecutive census years
}
\keyword{datasets}
