% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{population}
\alias{population}
\title{sample population (denominator) data}
\format{A matrix}
\usage{
population
}
\description{
Simulated denominator data with missing observation
}
\keyword{datasets}
