% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_akStats.R
\name{print_akStats}
\alias{print_akStats}
\title{Descriptive (Change) statistics}
\usage{
print_akStats(
  akObject,
  k = 3,
  reference = 1,
  N.quant = 4,
  show_Plots = TRUE,
  type = "lines",
  y.scaling = "fixed"
)
}
\arguments{
\item{akObject}{An output of \code{\link{akClustr}} function.
The object contains individual trajectories and their cluster
solution(s) at the specified values of \code{k}. Also, includes
the optimal value of \code{k} based on the criterion specified.
at (different) values of \code{k} the \code{traj}.}

\item{k}{[integer] \code{k} cluster to generate its solution.}

\item{reference}{[numeric] Specifying the reference line from
which the direction of each group is measured. Options are:
\code{1}: slope of mean trajectory, \code{2}: slope of medoid
trajectory, \code{3}: slope of a horizontal line
(i.e. slope = 0). Default: \code{1}.}

\item{N.quant}{[numeric] Number of equal intervals (quantiles)
to create between the reference line \code{(R)} and the medoids
\code{(M)} of the most-diverging groups of both sides of
\code{(R)}. Default is \code{4} - meaning quartile subdivisions
on each side of \code{(R)}. In this scenario, the function
returns the quartile in which the medoid of each group falls.
This result can be used to further categorize the groups into
'classes'. For example, groups that fall within the \code{1st}
quartile may be classified as 'Stable' groups (Adepeju et al. 2019).}

\item{show_Plots}{[TRUE or FALSE] To display cluster plots.
Defaults \code{TRUE}}

\item{type}{[character] plot type. Available options are:
\code{"lines"} and \code{"stacked"}.}

\item{y.scaling}{[character] works only if \code{type="lines"}.
\code{y.scaling} set the vertical scales of the cluster panels.
Options are: \code{"fixed"}: uses uniform scale for all panels,
\code{"free"}: uses variable scales for panels.}
}
\value{
A plot showing group membership or sizes (proportion)
and statistics.
}
\description{
This function perform two tasks:
(i) it generate the descriptive and change statistics
of groups, particularly suited for the outputs form
the \code{\link{akClustr}} function, and
(ii) generates the plots of the groups (performances).
}
\details{
Generates the descriptive and change statistics
of the trajectory groupings. Given an \code{akObject} class (from
the \code{akClustr} function), this function generates both the
descriptive and change statistics of the clusters.
There is also an option to show the plots of cluster groups.
The function draw from the functionalities of the
\code{ggplot2} library.
For a more customized visualisation, we recommend that users
deploy \code{ggplot2} directly (\code{Wickham H. (2016)}).
}
\examples{

data(traj)

trajectry <- dataImputation(traj, id_field = TRUE, method = 1,
replace_with = 1, fill_zeros = FALSE)

print(trajectry$CompleteData)

trajectry <- props(trajectry$CompleteData, id_field = TRUE)

akSolution <- akClustr(trajectry, id_field = TRUE,
method = "linear", k = c(3,5), crit='Calinski_Harabasz')

print_akStats(akSolution, k = 4, type="lines",
y.scaling="fixed", show_Plots = TRUE)

print_akStats(akSolution, k = 4, reference = 1,
show_Plots = TRUE, N.quant = 4, type="stacked")

}
\references{
\code{1}. Adepeju, M. et al. (2019). Anchored k-medoids:
A novel adaptation of k-means further refined to measure
inequality in the exposure to crime across micro places (Submitted).

\code{Wickham H. (2016). Elegant graphics for
Data Analysis. Spring-Verlag New York (2016)}
}
