% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{props}
\alias{props}
\title{Convert counts or rates to proportion}
\usage{
props(traj, id_field = FALSE)
}
\arguments{
\item{traj}{[matrix (numeric)]: longitudinal data. Each row represents an individual trajectory (of observations). The columns show the observations at consecutive time steps.}

\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.}
}
\value{
A matrix of proportion measures
}
\description{
This function converts counts or rates to proportion.
}
\details{
Given a matrix of observations (counts or rates), this function convert each observation to a proportion equivalent of the sum of each column. In other words, each observation is dividing by the sum of the column where it is located. , i.e. \code{prop = [a cell value] / sum[corresponding column]}
}
\examples{
traj <- gm.crime.sample1
head(traj) #
traj <- dataImputation(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
traj <- props(traj, id_field = TRUE)
print(traj)
}
