% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.R
\docType{data}
\name{simulated}
\alias{simulated}
\title{Simulated longitudinal dataset}
\format{
A data.frame
}
\usage{
simulated
}
\description{
Simulated longitudinal datasets containing
pre-defined clusters.
}
\keyword{datasets}
