% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TO1Risk.R
\docType{data}
\name{TO1Risk}
\alias{TO1Risk}
\title{Time-at-risk for the Adjudicated Toronto
Youth Data (Sample 1)}
\format{
A matrix
}
\usage{
TO1Risk
}
\description{
Time-at-risk per year for 378 individuals
in the Toronto, Ontario, Canada.
}
\references{
Nielsen, J. (2018) crimCV:
Group-Based Modelling of Longitudinal Data.
R package version 0.9.6.
URL https://CRAN.R-project.org/package=crimCV.
}
\keyword{datasets}
