% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj.R
\docType{data}
\name{traj}
\alias{traj}
\title{Longitudinal dataset}
\format{A matrix}
\usage{
traj
}
\description{
Simulated longitudinal datasets with missing values (\code{NA}, \code{Inf}, \code{null})
}
\keyword{datasets}
