% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traj_w_spaces}
\alias{traj_w_spaces}
\title{Sample longitudinal dataset containing whitespaces}
\format{
A dataframe with the following variables:
  \itemize{
  \item location_ids: Character id of sample locations
  at which values are obtained.
  \item X2001: Values at time step 1 (i.e. year 2001)
  \item X2002: Values at time step 2 (i.e. year 2002)
  \item X2003: Values at time step 3 (i.e. year 2003)
  \item X2004: Values at time step 4 (i.e. year 2004)
  \item X2005: Values at time step 5 (i.e. year 2005)
  \item X2006: Values at time step 6 (i.e. year 2006)
  \item X2007: Values at time step 7 (i.e. year 2007)
  \item X2008: Values at time step 8 (i.e. year 2008)
  \item X2009: Values at time step 9 (i.e. year 2009)
    }
}
\usage{
traj_w_spaces
}
\description{
Longitudinal dataset with both trailing and leading
whitespaces. For example, there is a trailing whitespace
at cell [3, 6], while there is a leading whitespace
at cell [9, 4].
}
\keyword{datasets}
