% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_wSpaces.R
\docType{data}
\name{traj_wSpaces}
\alias{traj_wSpaces}
\title{Datasets with white spaces}
\format{
A dataframe
}
\usage{
traj_wSpaces
}
\description{
Trajectories containing white spaces
}
\keyword{datasets}
