% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wSpaces.R
\name{wSpaces}
\alias{wSpaces}
\title{Removing whitespaces}
\usage{
wSpaces(traj)
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observation at consecutive time steps.}
}
\value{
A cleaned data.frame with whitespaces (if any) removed.
}
\description{
This function removes all leading and trailing whitespaces in longitudinal data
}
\examples{
traj <- gm.crime.sample1
wSpaces(traj)
}
\references{
\url{https://en.wikipedia.org/wiki/Whitespace_character}
}
