% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akmedoids.clust.R
\name{akmedoids.clust}
\alias{akmedoids.clust}
\title{Clustering of longitudinal data}
\usage{
akmedoids.clust(traj, id_field = FALSE, method = "linear", k = c(3,6)) ##' #details Given a list of trajectories represented in a matrix or data.frame, and a method for choosing initial cluster centroids (e.g. \code{\link{lpm.centroids}}), a list of clusters is generated after a limited number of iterations.
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observations at consecutive time steps.}

\item{id_field}{Whether the first column is a unique (\code{id}) field. Default: \code{FALSE}}

\item{method}{Initialisation strategy. Available method: \code{linear}}

\item{k}{either an exact integer number of clusters, or a vector of length two specifying the minimum and maximum numbers of clusters to be examined. The default is \code{c(3,15)}. When k is a range, the actual number of clusters is determined by Calinski-Harabatz criterion. number of clusters to generate. Default: \code{3}: (minimum value)}
}
\value{
A list containing cluster solutions for all value of k, including the solution at the optimal value of \code{k}, based on the Calinski-Harabatz criterion \code{(Calinski T, Harabasz J, 1974)}
}
\description{
This function group trajectories based on a functional definition of their long-term trends.
}
\examples{
traj <- gm.crime.sample1
print(traj)
traj <- missingVal(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
traj <- props(traj, id_field = TRUE)
print(traj)
output <- akmedoids.clust(traj, id_field = TRUE, method = "linear", k = c(3,6))
print(output)
}
\references{
\code{Calinski T, Harabasz J (1974) A dendrite method for cluster analysis. Commun Stat 3:1-27}
}
