% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akmedoids.clust.R
\name{akmedoids.clust}
\alias{akmedoids.clust}
\title{Clustering of longitudinal data}
\usage{
akmedoids.clust(traj, id_field = FALSE, k = c(3,6)) ##' #details Given a list of trajectories represented in a matrix or data.frame, and a method for choosing initial cluster centroids (e.g. \code{\link{lpm.centroids}}), a list of clusters is generated after a limited number of iterations.
}
\arguments{
\item{traj}{A matrix or data.frame with each row representing the trajectory of observations of a unique location. The columns show the observations at consecutive time steps.}

\item{id_field}{Whether the first column is a unique (\code{id}) field. Default: \code{FALSE}}

\item{k}{either an exact integer number of clusters, or a vector of length two specifying the minimum and maximum numbers of clusters to be examined. The default is \code{c(3,15)}. When k is a range, the actual number of clusters is determined by Calinski-Harabatz criterion. number of clusters to generate. Default: \code{3}: (minimum value)}
}
\value{
The original (\code{traj}) data with cluster label appended
}
\description{
This function group trajectories based on a given list of initial centroids
}
\examples{
traj <- gm.crime.sample1
print(traj)
traj <- missingVal(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
print(traj)
clusters <- akmedoids.clust(traj, id_field = FALSE, k = c(3,6))
print(clusters)
}
\references{
\code{Christophe Genolini, Xavier Alacoque, Marianne Sentenac, Catherine Arnaud (2015). kml and kml3d: R Packages to Cluster Longitudinal Data. Journal of Statistical Software, 65(4), 1-34. URL http://www.jstatsoft.org/v65/i04/}
}
