% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotr.R
\name{plotr}
\alias{plotr}
\title{Plotting the cluster solution}
\usage{
plotr(clustr, traj, id_field = TRUE, y.scaling = "fixed")
}
\arguments{
\item{clustr}{A vector of cluster labels (e.g. from an output of \code{\link{akmedoids.clust}})}

\item{traj}{The matrix or data.frame of trajectories from which the cluster solution \code{clustr} is generated.}

\item{id_field}{Whether the first column of \code{traj} is a unique (id) field. Default: \code{TRUE}.}

\item{y.scaling}{To set the vertical scales of the cluster panels. Options are: \code{"fixed"}: uses uniform scale for all panels, \code{"free"}: uses variable scales for panels.}
}
\value{
A plot showing clusters of trajectories, including the mean trajectory of each group.
}
\description{
Function to plot the clusters generated by the \code{\link{akmedoids.clust}} function
}
\details{
Given a matrix or data.frame with some suspected outlier observations, this function identified those observations based on the method chosen and replace all the observations accordingly.
}
\examples{
traj <- gm.crime.sample1
print(traj)
traj <- missingVal(traj, id_field = TRUE, method = 2, replace_with = 1,
fill_zeros = FALSE) #filling the missing values
print(traj)
traj <- props(traj, id_field = TRUE)
clustr <- akmedoids.clust(traj, id_field = TRUE, k = c(3,6))
print(plotr(clustr, traj, id_field=TRUE, y.scaling="fixed"))
}
\references{
\code{Wickham H. (2016). Elegant graphics for Data Analysis. Spring-Verlag New York (2016)}
}
\keyword{clusters}
\keyword{plot,}
