% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated}
\alias{simulated}
\title{Simulated longitudinal dataset}
\format{
A dataframe with the following variables:
\itemize{
  \item X1: Values across locations at time step 1
  \item X2: Values across locations at time step 2
  \item X3: Values across locations at time step 3
  \item X4: Values across locations at time step 4
  \item X5: Values across locations at time step 5
  \item X6: Values across locations at time step 6
  \item X7: Values across locations at time step 7
  \item X8: Values across locations at time step 8
  \item X9: Values across locations at time step 9
  \item X10: Values across locations at time step 10
  \item X11: Values across locations at time step 11
  \item X12: Values across locations at time step 12
  \item X13: Values across locations at time step 13
  \item X14: Values across locations at time step 14
  \item X15: Values across locations at time step 15
  \item X16: Values across locations at time step 16
  \item X17: Values across locations at time step 17
  \item X18: Values across locations at time step 18
  \item X19: Values across locations at time step 19
  \item X20: Values across locations at time step 20
  \item X21: Values across locations at time step 21
    }
}
\usage{
simulated
}
\description{
Contains simulated trajectories belonging to one of the
three pre-defined groups, namely (a) decreasing, (b) stable
and (c) increasing groups.
}
\keyword{datasets}
