% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TO1Risk}
\alias{TO1Risk}
\title{Time-at-risk for the Adjudicated Toronto
Youth Data (Sample 1)}
\format{
A dataframe with the following variables:
  \itemize{
  \item 8: Time-at-risk per year at age 8
  \item 9: Time-at-risk per year at age 9
  \item 10: Time-at-risk per year at age 10
  \item 11: Time-at-risk per year at age 11
  \item 12: Time-at-risk per year at age 12
  \item 13: Time-at-risk per year at age 13
  \item 14: Time-at-risk per year at age 14
  \item 15: Time-at-risk per year at age 15
  \item 16: Time-at-risk per year at age 16
  \item 17: Time-at-risk per year at age 17
  \item 18: Time-at-risk per year at age 18
  \item 19: Time-at-risk per year at age 19
  \item 20: Time-at-risk per year at age 20
  \item 21: Time-at-risk per year at age 21
  \item 22: Time-at-risk per year at age 22
  \item 23: Time-at-risk per year at age 23
  \item 24: Time-at-risk per year at age 24
  \item 25: Time-at-risk per year at age 25
  \item 26: Time-at-risk per year at age 26
  \item 27: Time-at-risk per year at age 27
  \item 28: Time-at-risk per year at age 28
  \item 29: Time-at-risk per year at age 29
  \item 30: Time-at-risk per year at age 30
  \item 31: Time-at-risk per year at age 31
  \item 32: Time-at-risk per year at age 32
  \item 33: Time-at-risk per year at age 33
  \item 34: Time-at-risk per year at age 34
  \item 35: Time-at-risk per year at age 35
  \item 36: Time-at-risk per year at age 36
  \item 37: Time-at-risk per year at age 37
  \item 38: Time-at-risk per year at age 38
    }
}
\usage{
TO1Risk
}
\description{
Real-life time-at-risk per year for 378 individuals
from the age of 8 to 38 in the Toronto, Ontario, Canada.
The data is obtained through the R package `crimCV`.
For further information, please see: Nielsen, J. (2018)
crimCV: Group-Based Modelling of Longitudinal Data.
R package version 0.9.6.
URL https://CRAN.R-project.org/package=crimCV.
}
\keyword{datasets}
