% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wSpaces.R
\name{wSpaces}
\alias{wSpaces}
\title{Removing whitespaces}
\usage{
wSpaces(traj)
}
\arguments{
\item{traj}{[matrix (numeric)]: longitudinal data. Each row represents an individual trajectory (of observations). The columns show the observations at consecutive time steps.}
}
\value{
A matrix with all whitespaces (if any) removed.
}
\description{
This function removes all the leading and the trailing whitespaces in a longitudinal data
}
\details{
Given a matrix suspected to contain whitespaces, this function removes all the whitespaces and returns a cleaned data. ’Whitespaces’ are white characters often introduced into data during data entry or data collection by recording devices. For example, neither " A" nor "A " equates "A" because of the whitespaces that exist in them.
}
\examples{
traj <- gm.crime.sample1
wSpaces(traj)
}
\references{
\url{https://en.wikipedia.org/wiki/Whitespace_character}
}
